urgh that's supposed to be an ulong

+++ Dimension/Model/Commands/HelloCommand.cs
@@ -11,7 +11,7 @@ namespace Dimension.Model.Commands
         public ulong id;
         public ulong myShare;
         public string username;
-        public Dictionary<int, int> peerCount;
+        public Dictionary<ulong, int> peerCount;
         public int externalControlPort;
         public int externalDataPort;
         public int internalControlPort;
